NXP-25168: refactoring to remove nuxeo-core dependency in nuxeo-core-io
